Remove redundant code (see full commit message)
This code was brought in based on the idea of "resetting" stack state which doesn't go back into its original "new" state.  That is, setting the next index as the sentinel value (-1)....Even though it wouldn't contribute to the correctness of the other functions.

For an empty stack, it doesn't matter what the starting value of `top_index_next` is because when we push an element onto an empty stack, the `top_index_next` will be created based off the value of `top_index`.  So, all that matters is that `top_index` is reset correctly, for both `top_index_next` and `top_index` to work.

This means that the code I added was redundant.  My thinking was too narrow while writing that.
The solution will be to remove code related to resetting `top_index_next`.